Switch to regparm(3) on x86_32
The old trigger code was unintentionally creating register pressure by
forcing the parameters into registers that were NOT clobbered as part
of a cdecl call. This is counter to out intentions: we want to use
EAX, EDX, and ECX whenever possible since those don't have to be saved
by the callee.

As a nice side effect, the switch to regparm(3) also allows us to
remove most of the stack setup/cleanup code.